move client.go to util directory
